+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v0.6.1
+
+- Fix the allowlist on macOS (#38).
+- Add fallback cover search for albums with ` - EP` and ` - Single`  suffixes added by Apple Music (#39).
+- Minor changes to how the player ID and name are determined (mainly on macOS).
+
 ## v0.6.0
 
 - Ported to MacOS.